half fix compile error

+++ dosbox/src/hardware/dma.cpp
@@ -544,7 +544,7 @@ Bit8u POD_State_Find_DMA_Callback( Bit32u addr )
 
 	size = sizeof(dma_state_callback_table) / sizeof(Bit32u);
 	for( int lcv=0; lcv<size; lcv++ ) {
-		if( (Bit32u) dma_state_callback_table[lcv] == addr ) return lcv;
+		if( (Bit32u) (uintptr_t) dma_state_callback_table[lcv] == addr ) return lcv;
 	}
 
 
@@ -555,7 +555,7 @@ Bit8u POD_State_Find_DMA_Callback( Bit32u addr )
 
 Bit32u POD_State_Index_DMA_Callback( Bit8u index )
 {
-	return (Bit32u) dma_state_callback_table[index];
+	return (Bit32u) (uintptr_t) dma_state_callback_table[index];
 }
 
 
@@ -564,7 +564,7 @@ void DmaChannel::SaveState( std::ostream& stream )
 	Bit8u dma_callback;
 
 
-	dma_callback = POD_State_Find_DMA_Callback( (Bit32u) (callback) );
+	dma_callback = POD_State_Find_DMA_Callback( (Bit32u) (uintptr_t) (callback) );
 
 	//******************************************
 	//******************************************
@@ -600,7 +600,7 @@ void DmaChannel::LoadState( std::istream& stream )
 	Bit8u dma_callback;
 
 
-	dma_callback = POD_State_Find_DMA_Callback( (Bit32u) (callback) );
+	dma_callback = POD_State_Find_DMA_Callback( (Bit32u) (uintptr_t) (callback) );
 
 	//******************************************
 	//******************************************
